Conceal the AD join password using an ENV var
Without this patch the AD join password is shown in the Puppet logs:

    Aug 15 16:42:10 agent1 puppet-agent[2476]: \
      (/Stage[main]/Realmd::Join::Password/Exec[realm_join_with_password]/returns) \
      change from notrun to 0 failed: echo  Password1' | realm join puppet.vm \
      --unattended --user=Administrator returned 1 instead of one of [0]

This patch addresses the problem by wrapping the redirection in a shell script
and reading the password from an environment variable set by Puppet.  This
prevents the password from showing up in the Puppet logs or in the process
table.

Resolves #9